requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
distlib==0.3.8
Django==5.0.4
django-bootstrap4==24.3
django-bootstrap5==24.2
django-cors-headers==4.3.1
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==2.1
djangorestframework==3.15.1
filelock==3.14.0
mysqlclient==2.2.4
pillow==10.3.0
platformdirs==4.2.1
PyMySQL==1.1.0
setuptools==69.5.1
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.1
